+++ custom_components/sonoff/switch.py
@@ -155,6 +155,24 @@ class XIntSwitch(XEntity, SwitchEntity):
         await self.ewelink.send_cloud(self.device, {self.param: 0})
 
 
+class XAutoControlSwitch(XIntSwitch):
+    """THR3 automatic-control toggle with a proven LAN endpoint."""
+
+    async def async_turn_on(self, **kwargs):
+        await self.ewelink.send(
+            self.device,
+            {self.param: 1},
+            cmd_lan="autoControlEnabled",
+        )
+
+    async def async_turn_off(self):
+        await self.ewelink.send(
+            self.device,
+            {self.param: 0},
+            cmd_lan="autoControlEnabled",
+        )
+
+
 class XBoolSwitch(XEntity, SwitchEntity):
     def set_state(self, params: dict):
         self._attr_is_on = params[self.param]